Add a kludge to get workaround a bug in the pytypes pre-commit wrapper.
The pre-commit pytypes wrapper we're using has a bug.  When only
unittest files are in the commit, pytype is invoked with no files:
https://github.com/mattseymour/pre-commit-pytype/issues/4

Workaround this for now by adding the suds directory to the arguments
for pytype.  This is a bit to heavyweight to be a good general solution
so hopefully the author will find a better way (or I could add a pytpe
hook to captain_hooks)