Implement a SudokuCell class
Decided that we're going to store information on potential values that
a cell can take and remain a valid sudoku into the cell itself.  So
we're going to need a cell object, not just an integer.

The currently implemented Sudoku class stores a set of potential values
which the cell could take on internally.  When the list of values is
reduced to a single one, a property will say that the sudoku cell has
been solved and another property will return what the cell's value is.

A single value can be set (useful for initial setup of the board
or for strategies which can determine what the cell's value should be
without eliminating all the other possibilities) and removing
possibilities is done via the `-=` operator.